refactor the configuration directive (and source file structure) so that the memcached configuration can be shared with session cache and memcached-based session ticket store
