+++ t/00unit/test.h
@@ -65,7 +65,7 @@ void test_lib__handler__mimemap_c(void);
 void test_lib__handler__redirect_c(void);
 void test_lib__http2__hpack(void);
 void test_lib__http2__scheduler(void);
-void test_src__ssl_session_ticket_c(void);
+void test_src__ssl_c(void);
 void test_issues293(void);
 
 #endif
