+++ t/40session-ticket.t
@@ -7,35 +7,54 @@ use t::Util;
 
 plan skip_all => "could not find openssl"
     unless prog_exists("openssl");
-plan skip_all => "openssl 1.0.2 or above is required"
-    unless `openssl version` =~ /^OpenSSL 1\.(?:0\.[2-9][^0-9]|[1-9])/s;
+#plan skip_all => "openssl 1.0.2 or above is required"
+#    unless `openssl version` =~ /^OpenSSL 1\.(?:0\.[2-9][^0-9]|[1-9])/s;
 
 my $tempdir = tempdir(CLEANUP => 1);
 
-my $tickets_file = ASSETS_DIR . "/session_tickets.yaml";
-
-subtest "reuse" => sub {
-    spawn_and_connect("file", $tickets_file, "New");
-    spawn_and_connect("file", $tickets_file, "Reused");
-    spawn_and_connect("file", $tickets_file, "Reused");
-    spawn_and_connect("off", "/dev/null", "New");
+subtest "internal" => sub {
+    spawn_with("ticket", "internal", "unused", sub {
+        is test(), "New";
+        test(); # openssl 0.9.8 seems to return "New" (maybe because in the first run we did not specify -sess_in)
+        is test(), "Reused";
+        is test(), "Reused";
+    });
+    spawn_with("ticket", "internal", "unused", sub {
+        is test(), "New";
+    });
 };
 
-subtest "missing tickets" => sub {
-    spawn_and_connect("file", $tickets_file, "New");
-    spawn_and_connect("file", $tickets_file, "Reused");
-    spawn_and_connect("file", "/dev/null", "New");
+subtest "file" => sub {
+    my $tickets_file = "t/40session-ticket/forever_ticket.yaml";
+    spawn_with("ticket", "file", $tickets_file, sub {
+        is test(), "New";
+        is test(), "Reused";
+        is test(), "Reused";
+    });
+    spawn_with("ticket", "file", $tickets_file, sub {
+        is test(), "Reused";
+    });
+};
+
+subtest "no-tickets-in-file" => sub {
+    my $tickets_file = "t/40session-ticket/nonexistent";
+    spawn_with("ticket", "file", $tickets_file, sub {
+        is test(), "New";
+        is test(), "New";
+        is test(), "New";
+    });
 };
 
 done_testing;
 
-sub spawn_and_connect {
-    my ($mode, $tickets_file, $expected) = @_;
-    my $server = spawn_h2o(<< "EOT");
+my $server;
+
+sub spawn_with {
+    my ($mode, $store, $tickets_file, $cb) = @_;
+    $server = spawn_h2o(<< "EOT");
 ssl-session-resumption:
-  mode: off
-ssl-session-ticket:
   mode: $mode
+  ticket-store: $store
   file: $tickets_file
 hosts:
   default:
@@ -43,6 +62,10 @@ hosts:
       /:
         file.dir: @{[ DOC_ROOT ]}
 EOT
+    $cb->();
+}
+
+sub test {
     my $lines = do {
         my $cmd_opts = (-e "$tempdir/session" ? "-sess_in $tempdir/session" : "") . " -sess_out $tempdir/session";
         open my $fh, "-|", "openssl s_client $cmd_opts -connect 127.0.0.1:$server->{tls_port} 2>&1 < /dev/null"
@@ -52,5 +75,5 @@ EOT
     };
     $lines =~ m{---\n(New|Reused),}s
         or die "failed to parse the output of s_client:{{{$lines}}}";
-    is $1, $expected;
+    $1;
 }
